segmentation mask for functional methods
NWB packaging failed if CaImAn or functional Suite2p are used instead of Cellpose, due to lacking `segmentation_mask`. This code generates the `segmentation_mask` from the ROIs by assigning each pixel to the ROI with the highest weight (handling overlaps), or the background if the pixel does not contain any ROIs.